Rename/renamespace AuthenticationConfiguration
