+++ tests/Application/AbstractCreateSerializedSuiteTest.php
@@ -11,7 +11,7 @@ use App\Enum\SerializedSuite\State;
 use App\Repository\SerializedSuiteRepository;
 use App\Repository\SourceRepository;
 use App\Repository\SuiteRepository;
-use App\Tests\Services\AuthenticationConfiguration;
+use App\Tests\Services\AuthenticationProvider\Provider;
 use App\Tests\Services\SourceOriginFactory;
 use App\Tests\Services\SuiteFactory;
 
@@ -41,10 +41,10 @@ abstract class AbstractCreateSerializedSuiteTest extends AbstractApplicationTest
     /**
      * @dataProvider serializeSuccessDataProvider
      *
-     * @param callable(AuthenticationConfiguration): SourceInterface $sourceCreator
-     * @param callable(SourceInterface): Suite                       $suiteCreator
-     * @param array<string, string>                                  $payload
-     * @param array<string, string>                                  $expectedResponseParameters
+     * @param callable(Provider): SourceInterface $sourceCreator
+     * @param callable(SourceInterface): Suite    $suiteCreator
+     * @param array<string, string>               $payload
+     * @param array<string, string>               $expectedResponseParameters
      */
     public function testSerializeSuccess(
         callable $sourceCreator,
@@ -87,7 +87,7 @@ abstract class AbstractCreateSerializedSuiteTest extends AbstractApplicationTest
     {
         return [
             'file, empty tests' => [
-                'sourceCreator' => function (AuthenticationConfiguration $authenticationConfiguration) {
+                'sourceCreator' => function (Provider $authenticationConfiguration) {
                     return SourceOriginFactory::create(
                         type: 'file',
                         userId: $authenticationConfiguration->getUser(self::USER_1_EMAIL)->id,
@@ -100,7 +100,7 @@ abstract class AbstractCreateSerializedSuiteTest extends AbstractApplicationTest
                 'expectedResponseParameters' => [],
             ],
             'file, non-empty tests' => [
-                'sourceCreator' => function (AuthenticationConfiguration $authenticationConfiguration) {
+                'sourceCreator' => function (Provider $authenticationConfiguration) {
                     return SourceOriginFactory::create(
                         type: 'file',
                         userId: $authenticationConfiguration->getUser(self::USER_1_EMAIL)->id,
@@ -113,7 +113,7 @@ abstract class AbstractCreateSerializedSuiteTest extends AbstractApplicationTest
                 'expectedResponseParameters' => [],
             ],
             'git' => [
-                'sourceCreator' => function (AuthenticationConfiguration $authenticationConfiguration) {
+                'sourceCreator' => function (Provider $authenticationConfiguration) {
                     return SourceOriginFactory::create(
                         type: 'git',
                         userId: $authenticationConfiguration->getUser(self::USER_1_EMAIL)->id,
@@ -126,7 +126,7 @@ abstract class AbstractCreateSerializedSuiteTest extends AbstractApplicationTest
                 'expectedResponseParameters' => [],
             ],
             'git with ref request parameters' => [
-                'sourceCreator' => function (AuthenticationConfiguration $authenticationConfiguration) {
+                'sourceCreator' => function (Provider $authenticationConfiguration) {
                     return SourceOriginFactory::create(
                         type: 'git',
                         userId: $authenticationConfiguration->getUser(self::USER_1_EMAIL)->id,
@@ -143,7 +143,7 @@ abstract class AbstractCreateSerializedSuiteTest extends AbstractApplicationTest
                 ],
             ],
             'git with request parameters including ref' => [
-                'sourceCreator' => function (AuthenticationConfiguration $authenticationConfiguration) {
+                'sourceCreator' => function (Provider $authenticationConfiguration) {
                     return SourceOriginFactory::create(
                         type: 'git',
                         userId: $authenticationConfiguration->getUser(self::USER_1_EMAIL)->id,
