+++ tests/Application/AbstractListSuitesTest.php
@@ -6,7 +6,7 @@ namespace App\Tests\Application;
 
 use App\Repository\SourceRepository;
 use App\Repository\SuiteRepository;
-use App\Tests\Services\AuthenticationConfiguration as AuthConfig;
+use App\Tests\Services\AuthenticationProvider\Provider as AuthConfig;
 use App\Tests\Services\SourceOriginFactory;
 use App\Tests\Services\SuiteFactory;
